[BugFix][Kernel] Prevent QLI k-scale address overflow (#16824)
### What this PR does / why we need it?

The paged arch22 QLI v2 kernel stores `kScaleBlkStride` as `int32_t`.
When the product of a physical block ID and the key dequantization scale
stride exceeds `INT32_MAX`, signed 32-bit arithmetic wraps and the
kernel reads K-scale data from an incorrect address. The resulting
candidate indices can be corrupted and may cause incorrect model output.

This patch changes `kScaleBlkStride` and its explicit conversion to
`int64_t`, so the physical page address is calculated using 64-bit
arithmetic.

With a scale stride of 73,856 FP16 elements, the first overflowing
physical block ID is 29,077.

This PR intentionally changes only the verified arch22 (A2/A3) path. The
arch35 (A5) path remains unchanged pending validation on A5 hardware.

### Does this PR introduce _any_ user-facing change?

Yes. It fixes occasional corrupted model output on A2/A3 when paged QLI
v2 caches allocate sufficiently high physical block IDs. There is no API
or configuration change.

### How was this patch tested?

An A2/A3 NPU nightly regression was added to
`tests/e2e/nightly/single_node/ops/singlecard_ops/test_quant_lightning_indexer_v2.py`.
It stores identical key and K-scale data at low physical pages and at
the first pages beyond the signed-int32 offset boundary, then requires
the Top-K indices to match exactly. A guard prefix makes the old wrapped
address deterministic. The case requires about 8 GiB and skips when less
than 10 GiB of NPU memory is free.

The patch was also tested on Ascend 910_9391 with CANN 9.1.0 using the
QLI v2 operator and a DeepSeek V4.1 W8A8 deployment.

- High-page operator reproducer, original binary: failed; high-page
Top-K cutoff differed from the low-page/CPU reference.
- Rebuilt binary with this patch: passed; high-page and low-page Top-K
results were identical.
- Original binary restored: failed again with the same mismatch.
- Real-model A/B/A case at physical block ID 29,198: the original
operator matched 0/512 CPU-reference Top-K candidates and produced
corrupted output; the patched operator matched 512/512 and produced
normal output; restoring the original operator reproduced the failure.
- Repeated the original six request cases 24 times with the patched
operator without observing corrupted output.
- `ruff check`, `ruff format --check`, Python syntax compilation, and
`git diff --check` pass locally.

- vLLM main:
https://github.com/vllm-project/vllm/commit/84030bbe3d74d99bad477a3d2e37a973ccd8865c

+++ csrc/attention/quant_lightning_indexer_v2/op_kernel/arch22/quant_lightning_indexer_v2_service_vector_arch22.h
@@ -165,8 +165,10 @@ __aicore__ inline void QLIV2Vector<QLIV2T>::GetKeyScale(const QLIV2Common::RunIn
     if constexpr (PAGE_ATTENTION) {
         // A11: k_scale 0 轴非连续 — 块基址 = 表值 x keyDequantScaleStride0 (真实 stride);
         // 0 时兜底原紧凑公式 (kCacheBlockSize_), 现网行为不变 (R8)
-        int32_t kScaleBlkStride = constInfo_.keyDequantScaleStride0 != 0 ?
-                                      static_cast<int32_t>(constInfo_.keyDequantScaleStride0) : kCacheBlockSize_;
+        // Paged hybrid caches can span more than 2^31 scale elements. Promote
+        // before multiplication so high physical block IDs do not wrap.
+        int64_t kScaleBlkStride = constInfo_.keyDequantScaleStride0 != 0 ?
+                                      static_cast<int64_t>(constInfo_.keyDequantScaleStride0) : kCacheBlockSize_;
         int32_t startBlockTableIdx = startS2 / kCacheBlockSize_;
         int32_t startBlockTableOffset = startS2 % kCacheBlockSize_;
         int32_t blockTableBatchOffset = batchId * maxBlockNumPerBatch_;
